added kconfig infrastructure, the trivial way
This comes from the "bst-kconfig" project, which lives at
  git://git.informatik.uni-erlangen.de/vamos/bst-kconfig

In particular, I copied from commit v2.6.34-4-g8168c0b which exported
to a standalone package the kconfig code from version 2.6.34 of the
kernel. This commit copies the files literally, but it renames
"Makefile" to "Makefile.kconfig", as our Makefile is already in place.

It can be said that it is a lazy choice and the code is old, but I
also evaluated other projects as a source of Kconfig code.  In
particular, I studied busybox and buildroot; but the work needed to
extract the relevant code looks bigger. I also evaluated
kconfig-frontends, but it is not what we needed here (besides, it uses
automake and thus it's completely unmaintainable for me -- known
hackers say that autotools must die, and I share the idea).

If needed, we can replace this with a newer Kconfig setup later on.
At the time being, this is code base is more than enough for this
project.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>