bugfix in Makefile: fix a build bug introduced by "three fixes"
Commit "5433fa2 Makefile: three fixes" introduced a bug, so that a
fresh clone fails to compile because of a missing "board.h" file.

This makes all objects depends on the symlink, which in turn is
removed by "make clean", so you don't need to start from a fresh clone
to experience the complete build.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>

Conflicts:

	Makefile