doc: trivial: aesthetics in source, no effect on output
I prefer the comment separatos for sections to be different than those
for chapters (so I better see the structure while editing). It makes
no difference for others I'm sure.

Also, remove the info menus: they tend to rust in the source
while the good ones in the output are regenarated by ./fixinfo .

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>