arch/lm32/ram.ld: remove all cruft
Remove all unneeded stuff from the linker script. Now the script only
includes what is actually used: it is easier to understand and change
(we'll add some local ELF sections), and it makes a better reference
for possible porting to a different architecture.

This commit also moves the default libraries from the linker script
itself to the LDFLAGS in Makefile, where they really belong.

CONFIG_DETERMINISTIC_BINARY confirms this makes no difference in
the generated binary (nor the ELF FWIW).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>