arch/lm32/ram.ld: provide mprintf if missing
When we'll enable pp_printf, we'll need to resolve mprintf to
it. Local files are fixed by #define in wrc.h, but ptp-noposix has
explicit mprintf calls inside.  By using "PROVIDE" in the linker
script, we resolve mprintf to pp_printf only if no native mprintf
is part of the link stage. This has no effect if pp_printf is
not selected.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>