Makefile: bugfix (my bug) for git submodules
The rule for "git submodule update" was wrong, so the command was
always performed.  This is a pain if you are using commits in the
submodule that are not yet committed in the supermodule. This always
happens during development.

I'm sorry for the inconvienence, and I thank Aurelio for noting this.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>