softpll/ copied back from ptp-noposix
After several back and forth, we agreed that softpll should live in
wrpc-sw, even if it is also used on the wr switch.  I don't pick up
the history, because the tool got renamed and moved several times.

This comes from commit 1e2c71d1f0e1094adc6c2eb46e30d4164262f85c, the
current master of ptp-noposix (the one that is currently a subproject)

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>