kill mprintf, since pp_printf is better and solid
mprintf is of uncertain copyright status, and that's why
we added pp_printf (actually, that's why I picked pp_printf
fro ptp-proposal and made it a standalone thing).

Now that everyone involved confirms pp_printf is well tested,
let's remove mprintf.  We still accept mprintf in the callers,
because the name is used by our ptp-noposix submodule, and I'm
sure people has it in their fingers and it will appear again.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>