Optionally detect reset and print a stack trace
The new configuation option CONFIG_CHECK_RESET adds a function to dump
the stack (to the serial port) if a reset is detected. This happens
when a null pointer is followed or when the program is reloaded
(spec-sw/tools/spec-cl, for example).  For this reason, the feature is
disabled by default (actually, spec-cl may clear the RAM first, but it
currently doesn't do that and the check is triggered).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>