Kconfig: ppsi_runtime_verbosity is enabled by default
The parameter used to be off by default in order to save space in the
generated binay.  Now ppsi is smaller and there's no reason to
prevent diagnostic code to be there.

This parameter, though, will disappear over time, as we have a new
diagnostic system, more modular (which still offers a compile-time
option to remove all diagnostic code).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>