Makefile: use bigobj.lds during initial partial link
This commit makes no effect, but it prepares for the following one.
In general, using an intermediate platform-agnostic linker script
when making wrc.o is not bad, so let's do it here.

The main reason why we need it is to force the "KEEP" command for the
".cmd" ELF section, introduced by the next commit.  The alternative to
this would be chaning the main linker script, which is worse in my
opinion.

If we add initcalls or similar stuff later, this bigobj.lds is the
place to do it.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>