lib: add usleep.c (and call its own init from main)
This adds a real usleep implementation. The init function calculates
the "loops per jiffy" variable, by watching the timer ticks before
interrupts are enabled by spll_init().  This init takes 40ms.

Then, the udelay runs the same loop used in calculating lpj, and
is thus correct (it will be slower because we have interrupts running
at run time, but this is not a problem).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>