dev: introduce uart-sw
This introduces dev/uart-sw as an alternative to dev/uart . The
software-only thing is a pair of circular buffers that can be driven
by tools/wrpc-uart-sw .

The software uart can be enabled by Kconfig (and is disabled by
default).  The code size is the same as the uart, but it "wastes" 306
bytes of data space.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>