tools: add wrpc-uart-sw.c
This is the host part of the software-only uart. The lm32 part is
selected at compile time (CONFIG_UART_SW).

The tool, when run, opens one or more pty pairs, and reports to stdout
which is the name of the slave descriptor. You can then run
"minicom -p /dev/pts/XXX" or other commands, to access the software uart.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>