lib/ptp-noposix-wrappers.c: copied from ptp-noposix
This is a copy of ptp-noposix/libposix/freestanding-wrapper.c, moved
into this project because it actually is only used here. The Makefile
now builds this one.

This move is needed in order to remove this faulty usleep and provide a
better one, while keeping the other fake functions called by ptp-noposix
code.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>