w1: files for new onewire implementation (not built)
This adds the files for an alternative to sockitowm. I called it "w1"
like in the Linux kernel, because "onewire" was already in use as file
name in this project.

The bus scanning code was a fun project of mine a few years ago, the
rest is what I did for my own crappy operating system in the last few
days.  This commit adds the files as modified to build in wrpc-sw, since
I have no "official" Bathos files yet to diff against.

This commit doesn't change the build, only adds the file, so actual
changes to existing code is all part of the next (smaller) commit.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>