Makefile: change to support gmake-3.82
Split a rule into two rules, to make make happy.

In previous versions of make it was acceptable to list one or more
explicit targets followed by one or more pattern targets in the same
rule and it worked "as expected". However, this was not documented as
acceptable and if you listed any explicit targets AFTER the pattern
targets, the entire rule would be mis-parsed. This release removes
this ability completely: make will generate an error message if you
mix explicit and pattern targets in the same rule. You must split
these rules into two rules: one for the pattern and one for the
explicit targets. The Linux kernel source has already been modified in
this way (in newer kernels).