wrc.h: define printf as pp_printf
Sometimes a printf happens in the code, and it eventually prevents the
wrc to link because it pick up the whole of libc.  While ptp-noposix
already has a definition of printf to mprintf, we don't have the
same for ppsi, so let's add it.

Currently, rxts_calibrator.c includes calls to printf, and didn't
build with CONFIG_PPSI until this commit.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>