general: improve sw-uart support (only for expert config).
This commits affect several files, to be able to support both uart and
uart_sw at the same time. My aim is sending verbose ppsi diagnostics
to the uart_sw while retaining the real uart (or vuart) for the
interactive user and wrpc status gui.

To this aim, I use weak functions to avoid hairy ifdef around the code.
Now both hw and sw uarts are initialized, and the missing init just
does nothing. Also, uart_sw_write_string is always defined, and
it is an alias of uart_write_string if no uart_sw is built.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>