for-ppsi: Makefile changes to please "make -j8"
We need to link ppsi/proto-standard/libstd.a explicitly, but we also
need a rule to generate it, or parallel compilation fails saying it
has no rule. The makefile depended on build order (my fault), as
building ppsi/ppsi.o, with its rule), also built libstd.a

BTW, this removes ppsi/arch-wrpc/libarch.a from the dependencies,
because it is already linked in making ppsi/ppsi.o (otoh the need for
an explicit mention of libstd.a can't easily be solved in ppsi
itself, so it remains here).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>