for-ppsi: pass DIAG_PUTS to ppsi, and use new ppsi revision
With this commit, when you select both the hardware and software uart,
ppsi diagnostics (controlled  by the "verbose" command) got to
the software uart, not to the hardware one. In this way you can run
the shell in a clean tty.

Unfortunately, frame dumps still reach the hardware uart, if frame
verbosity is greater than 1.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>