check-error: new file, to catch undesired function calls
We don't want to call the real printf, which brings in the whole of
newlib, but it sometimes happen. So let's diagnose the problem
immediately rather than letting users wonder what is wrong with the
size.

Similarly, if building under ppsi, ensure we are not calling libgcc
division functions: any 64-bit division my be done with __div64_32
offered by ppsi/lib (copied from the kernel).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>