remove unused wrc_extra_debug
The variable was used to mask a few diagnostic messages but was always
set to 1. Since such messages are actually few at this time, remove
the variable altogether.

The code, however, included a hardwired check for servo messages, used
by the ptp-noposix callers.  This commit preserves the behaviour while
removing the dependency from ptp-noposix headers.  The result is ugly
but temporary, because we are going to phase out ptp-noposix.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>