build: avoid internal "ptpdump" support in ppsi, when RAM won't fit
Etherbone won't fit with a full ppsi in the default RAM size for SPEC
cards. So this commit uses a ppsi hack to avoid internal ptpdump,
activating it when CONFIG_RAMSIZE is not 128kB and CONFIG_ETHERBONE is
set.

We should rather compare RAMSIZE with "< 131072", but neither gnu make
nor this version of Kconfig support numeric comparison.

This hack will soon disappera, as ppsi is gaining Kconfig support, so
wrpc-sw will be able to pass proper configuration. Also, 128kB will soon
be the default for SPEC images.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>