softpll-unify: header changes
softpll is built for both wrc and wrs; since it includes wrc.h and
syscon.h, we must make them suitable for both environments.

Waiting for the time and energies for a better merge, this uses ifdef
to differentiate the two cases: the timer code is different (different
prototypes) and the clock frequency is different as well (125 vs 62.5
MHz).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>