softpll-unify: add new files, from wr-switch-sw/rt
The commit is part of the effort in unifying softpll with
wr-switch-sw, and later remove the duplicated code there.

The files added by these commit are going to be built when configuring
wrpc-sw to build wr-switch rt_cpu (which is, basically, the softpll
code alone, with mini-rpc with the host and the basic glue code).  The
files are copied with the original name with two exceptions:

   wr-switch-sw::rt/main.c            becomes wrpc-sw::wrs_main.c
   wr-switch-sw::rt/arch/lm32/ram.ld  becomes wrpc-sw::arch/lm32/ram-wrs.ld

The files are copied from commit FIXME of wr-switch-sw.

This commit has no technical effect, as the files are not built for wrpc-sw.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>