Makefile: fix ram.ld dependencies
ram.ld must be rebuilt every time the configuration changes. It was
supposed to be like that, but it was not.

So, the older ram.ld was used when first building with a new config,
and only the second time the rule was used. This was exposed by
failuers in ./MAKEALL where each configuration is built only once.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>