dev/sdb-eeprom.c: register the storage tree with sdb-lib
It is not mandatory to actually register the tree, but it's
a good move for consistency.

Later we'll need to scan several devices, for example to look for a
mac address in carrier flash or eeprom. To do that we'll need to have
devices properly registered.

This costs 150 bytes of binary size in gsi_defconfig.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>