eeprom: rename "present" to "init"
eeprom_present() returned a value nobody used. Rename to eeprom_init()
instead. Also, save the two parameters so that they are not needed
in other functions (see later commits).

Also, turn int8_t to int as function arguments (int is better for the
CPU).

This adds 20 bytes to the legacy case, and removes 4 to the sdb case.
The legacy size is reduced by later commits.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>