softpll stats: remove valid bit
"char valid" in a structure of integer makes alignment problems
(because we are accessing it from a different CPU we'd better not have
holes).

Also, there is no inherent consistency needed here, and this valid bit
is racy anyways. We could solve it with a sequence number, if any, but
let's avoid it altogether by now.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>