ldscript and crt0.S: remove useless _gp
The global pointer is used by some special assembly instructions
that we are not using (i.e., the compiler is not generating them
for us, because we don't use PIC or other "strange" stuff.

So, we'd better remove an assignement in the linker script that makes
little sense (it must be a typo of some kind, in the dark ages of this
code base).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>