pfilter: name registers, don't use numbers
By naming registers the code is much more readable in my opinion.
And we avoid, in future modifications, to mistake a frame offset for
a register.

The enumeration of names starts from 1024 to deny use of simple
numbers. Also, the enumeration is verified to be exactly 32 names,
by raising a build error if not.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>