shell low-level commands, and use 0 as default delays
This adds two commands: devemem (to read and write registers/RAM)
and delays (to read and change the constant delays.  The commands are
only available for CONFIG_DEVELOPER builds.

The default delays were just wrong, they are valid of an older
gateware version.  Using 0, it is clearer to the user that the right
value is missing (if the SFP database is populated, the right value
applies anyways).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>