Makefile: simplify two rules
Simplify access to $CONFIG_RAMSIZE. Actually, the previous way
was not working for me, as my $SHELL is not bash.

The bug is uncovered by the next commit, because the "-s" argument
to genramvhd has always been wrong for me, but I didn't notice
(so maybe the "-s" argument has no effect? I can't tell at this point)

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>