stat: only print statistics when something happened
This change monitors the update_count of wr_servo, and only prints
a statistics line when something happens.

This means we won't have long "TRACK_PHASE" listings when the master
is dead. Moreover, we see changes when they happen, not at a random
time during the next second.

The bad side of the thing is that if nothing happens in the wr servo,
nothing is ever printed. But see next commit.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>