gui: print a new status screen whenever the wr servo runs
The gui is timely: it refreshes every that many seconds, but I'd
better see a change immediated when it happens.  This commit
introduces monitoring of the wr-servo update count, so if anything
changed an information scree is dumped.

This is different from stats (two commits ago), because we want to
refresh timely in any case, so the user sees the thing is active.

You can verify the effect of the commit after "refresh 4": timely
refresh is now slower than servo updated and you'll see how gui
updates are synced with the ptp protocol.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>