crt0.S: include information for the dumping tool
There was a hole at the beginning of the generated binary, so this
commit includes some data to identify the endianness of the dump file
and host pointers to the important structures that we dump.
For wr-switch, the ppi pointer is just set to zero.

The change is backward compatible, because we just fill 32 bytes of
data where we had a bunch of zeroes.

The next commit will use such information when available in the dump file.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>

SQUASH

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>