tool/wrpc-dump: change the access mode, to avoid "gcc -m32"
Accessing data structures in the host to get target fields is broken:
we may have different data size or alignment, or whatever. The hack
of using "-m32" to be able to access lm32 fields from x86-64 is not
clean (we imply the i386 and target have the same size/alignment).
Even worse, not everybody has an x86-64 cross-compiler for i386.

Thus, we use OFFSET_OF in the target, with the lm32 cross compiler,
to build a binary table that is then accessed by the host.

In order to link the data structure in the host we use the assembler
instead of objcopy. With objcopy we'd need to spell out the bfd name
of the host, and we'd get horrible names like
   struct dump_info _binary_dump_info_host_bin_start[];

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>