Revert "Makefile: add a depend rule, long missing"
This reverts commit 3a3c38065d681b5b27679226572c4c1452897f54.

The previous line: every object depending on .config, was better.
We had the problem that a fresh clone was not building because of a
missing autoconf.h.

The depend rule was not working anyways, because it created an empty
.depend the first time over, due to the missing autoconf.h.  I admit I
don't know how to create a depend file after configuration: make
pretends to remake all its inclusions as the first step, so a rule
like ".depend: silentoldconfig" loops forever.