Makefile: be quiet in calling Kbuild makefiles
This quiet=quiet_ has the effect of printing

  HOSTCC  scripts/kconfig/conf.o
  HOSTCC  scripts/kconfig/kxgettext.o
  SHIPPED scripts/kconfig/zconf.tab.c
  SHIPPED scripts/kconfig/lex.zconf.c
  SHIPPED scripts/kconfig/zconf.hash.c
  HOSTCC  scripts/kconfig/zconf.tab.o
  HOSTLD  scripts/kconfig/conf

instead of the full command lines. We could even use quite=silent_ to
have no output at all, but this would introduce a noticeable delay in
the build.

The aim of this change is making more visible the error "your wrpc is
not configured" at the first build. The full compilation of config
stuff makes the error less clear, and the initial error confuses
people (me included).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>