include changes, no difference in binaries
The makefile included trace.h for everyone, which in turns included
wrc.h.  It's better to include wrc.h straight on (we can't rely on all
files including it, for example pp_printf does not).

Meanwhile, some redundant includes are removed, and some are added
(I prefer wrc.h to be explicitly listed in the files that use it,
even if the command line set in Makefile already prepends it).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>