net: move to caller-based allocation
We removed the pool of sockets, to only keep a list of pointers.
The caller provides the socket (and the buffer).

This saves from 440 to 1570 bytes in the binary (more in the ones
with no CONFIG_IP, beause sockets are not used.

Now we can move to a caller-defined queue size, when I find the time and
willingness.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>