net: let the caller allocate queue size
I also change the prototype of the socket creation so the compile can
warn me where the locally-allocated socket must be changed.

The size saving is not huge, but this allows later creation of further
sockets with a smaller overhead (e.g.: the syslog socket that sends only
has no queue, and the snmp socket needs a very small buffer).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>