net: define wr_ethhdr and use it (no change)
We used to have "struct ethhdr" local in net.c, passing "void *hdr" to
minic.  We exports the structure to an header so to use it in
communicating with the minic.  Unfortunaltey, ethhdr is defined elsewhere
too, so we'd better rename to wr_ethhdr.

This is a step towards the introduction of vlans, but has no technical
effect by itself.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>