net: don't run bootp if IP address is static
This solves http://www.ohwr.org/issues/1218

We used to have a "needIP" variable, but we need three states:
if the address was set by bootp, we run bootp again when the link goes
up, but if it was a static assignment we do not (and keep the static one).

So I renamend needIP to a 3-valued ip_status. The rename ensured the
compiler found all users.  Also, the messages are now 3-headed, so
the user know whether the address is static or it comes from bootp.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>