net: link_status is now a global variable
All "tasks" must have the same prototype, so we'd better get
void in all of them at the time being (so we only change the nerwork
function).

Meanwhile, use value "0" for "link down", so it can work as a simple
enable flag for the IP polling code.

(note: we query the link status in wrc_ptp_update() too, and we should
rather refer to the global variable)

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>