shell: optionally define a build-time init command
This commit adds the possibility to build an init command in the
binary.  It can be useful to install a number of nodes with a special
configuration, without the need to reach the shell and/or eeprom of
each one.

This is useful, for example, to pass syslog configuration
or the (yet to be implemented) definition of multiple vlans for
etherbone and the wr fabric (frame classes 7 and 6 f the packet
filter). Moreover, I personally love to load a master or slave node
for testing without interacting with the node (or changing the flash
over and over).

This commit also makes reading of commands from storage an optional
feature, but only in the advanced/developer options. Thus, you can
build a wrpc that uses the build-time init commands, or the
flash-stored init commands, both or none of them.

Size effects:
     * no effect if you keep the defaults (no build-time init, flash init)
     * +248 bytes (+ the command itself) if you enable build-time init
     * -1696 bytes if you remove flash-based init (discouraged)

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>