pfilter: novlan: non need to split ip_uni and ip_ok rules
This shortens the rule-set by one rule, making space for pdelay.

IP is acceptable whether unicast or broadcast. Thus, remove the IP_UNI
intermediate calculation and use only IP_OK.

Most of the lines in this commit are a renaming of FRAME_OUR_MAC to
FRAME_MAC_OK, with no real effect, because we'll coalesce all valid
macs in a single bit (I planned to use it in this commit, but I did
not, still I don't want to split the patch).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>